backend/pyproject.toml
[project]
name = "hybrid-rag-interviewer-backend"
version = "0.1.0"
description = "Intelligent Hybrid RAG Interviewer backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.1.19",
  "alembic>=1.13.1",
  "pydantic-settings>=2.3.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "redis>=5.0.4",
  "qdrant-client>=1.9.1",
  "sentence-transformers>=3.0.1",
  "numpy==1.26.4",
  "pypdf>=4.2.0",
  "python-multipart>=0.0.9",
  "httpx>=0.27.0",
  "orjson>=3.10.3",
  "insightface>=0.7.3",
  "onnxruntime==1.18.0",
  "opencv-python-headless==4.9.0.80"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.2",
  "pytest-asyncio>=0.23.7",
  "ruff>=0.4.9",
  "mypy>=1.10.0"
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
